Split llm helpers module
